perf: :zap: reduce inline php + adjust indentation + add comments
reduce inline php + adjust indentation + add comments to fit better with rubric + neater
