+++ index.php
@@ -1,12 +1,4 @@
-<?php require './php/mall_page.php';
-if (empty(session_id())) session_start();
-
-error_reporting(E_ERROR | E_PARSE);
-if (fopen('./php/install.php', 'r') != null) {
-  exit("'install.php' still exists! Delete it to proceed!");
-}
-
-?>
+<?php require './php/mall_page.php';?>
 <!DOCTYPE html>
 <html lang="en">
