+++ templates/about.php
@@ -1,15 +1,12 @@
 <?php
-if ( empty(session_id()) ) session_start();
-error_reporting(E_ERROR | E_PARSE);
+    // check session status and start session 
+    if (empty(session_id())) session_start();
 
-if (fopen('../php/install.php', 'r') != null) {
-    exit("'install.php' still exists! Delete it to proceed!");
-}
-// echo '<h2>$_SESSION values</h2>';
-// echo '<pre>';
-// print_r($_SESSION);
-// echo '</pre>';
-// echo '<hr>';
+    // detect install.php
+    error_reporting(E_ERROR | E_PARSE);
+    if (fopen('../../php/install.php', 'r') != null) {
+        exit("'install.php' still exists! Delete it to proceed!");
+    }
 ?>
 
 <!DOCTYPE html>
