--- templates/account/past_transaction.html
@@ -1,191 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Past Transactions</title>
-    <!-- Link CSS -->
-    <link rel="stylesheet" href="../../css/style.css" />
-    <link rel="stylesheet" href="../../css/account/account.module.css" />
-    <link
-      rel="stylesheet"
-      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css"
-      integrity="sha512-iBBXm8fW90+nuLcSKlbmrPcLa0OT92xO1BIsZ+ywDWZCvqsWgccV3gFoRBv0z+8dLJgyAHIhR35VZc2oM/gI1w=="
-      crossorigin="anonymous"
-    />
-  </head>
-
-  <body>
-    <!--Navigation bar-->
-    <header>
-      <!-- Logo -->
-      <div class="brand">
-        <a href="../../index.php"
-          ><img
-            src="https://i.imgur.com/mE6aWmB.png"
-            alt="logo"
-            class="logo-img"
-          />
-        </a>
-      </div>
-      <!-- Right menu -->
-      <nav class="menu">
-        <input type="checkbox" id="menuToggle" />
-        <label for="menuToggle" class="menu-icon"
-          ><i class="fa fa-bars"></i
-        ></label>
-        <ul>
-          <a href="../about.php">
-            <li>About us</li>
-          </a>
-          <a href="../fees.html">
-            <li>Fees</li>
-          </a>
-          <a href="../account/account.html">
-            <li>Account</li>
-          </a>
-          <a href="../browse-menu.html">
-            <li>Browse</li>
-          </a>
-          <a href="../faq.html">
-            <li>FAQs</li>
-          </a>
-          <a href="../contact.html">
-            <li>Contact</li>
-          </a>
-          <a href="../login-form.php">
-            <li>Sign in</li>
-          </a>
-          <a href="../cart.html" id="cart">
-            <li>Cart</li>
-          </a>
-        </ul>
-      </nav>
-    </header>
-    <!-- End header -->
-    <!--body part-->
-    <div class="body_container">
-      <div>
-        <p>Homepage/<a href="account.html">My account</a></p>
-        <br />
-        <h1>My account</h1>
-        <div class="account_body">
-          <!--left nav bar-->
-          <div class="account_nav">
-            <div class="account_nav_item">
-              <a href="account.html">My details</a>
-            </div>
-            <div class="account_nav_current_state">
-              <a href="past_transaction.html">Past transaction</a>
-            </div>
-          </div>
-          <!--the right main content-->
-          <div class="account_main_content">
-            <h1>Past transaction</h1>
-            <!--the past transaction-->
-            <div class="past_transaction" style="border-bottom: 0px">
-              <!--picture of the product-->
-              <div>
-                <img
-                  src="https://i.imgur.com/EVvlj0R.jpg"
-                  class="avatar"
-                  alt="product one"
-                />
-              </div>
-              <!--description of the transaction-->
-              <div class="product_description">
-                <div class="product">
-                  <h4>Homepod mini</h4>
-                  <p>Description: The transaction was made in 19/06/2020</p>
-                  <p>Delivery date: 23/06/2002</p>
-                  <p>Price:$99</p>
-                  <p>Amount: 2</p>
-                  <div class="confirmation">
-                    <p>Confirm order<i class="fas fa-check-circle tick"></i></p>
-                  </div>
-                </div>
-              </div>
-              <!--product picture-->
-              <div>
-                <img
-                  src="https://i.imgur.com/6KqDWVE.jpg"
-                  class="avatar"
-                  alt="product two"
-                />
-              </div>
-              <!--description of the transaction-->
-              <div class="product_description">
-                <div class="product" style="border-bottom: 0px">
-                  <h4>Apple watch series 6</h4>
-                  <p>Description: The transaction was made in 19/06/2020</p>
-                  <p>Delivery date: 24/06/2002</p>
-                  <p>Price: $399</p>
-                  <p>Amount: 1</p>
-                  <div class="confirmation">
-                    <p>Confirm order<i class="fas fa-check-circle tick"></i></p>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <!-- Footer -->
-    <footer class="page-footer">
-      <div class="container">
-        <div class="grid-container">
-          <!-- Footer Logo -->
-          <div class="grid-item">
-            <a href="../../index.php"
-              ><img
-                src="https://i.imgur.com/mE6aWmB.png"
-                alt="logo"
-                class="logo-img"
-            /></a>
-          </div>
-          <!-- Quick Link -->
-          <div class="grid-item inner-grid-container">
-            <div class="grid-item">
-              <a href="../about.php">About Us</a>
-            </div>
-            <div class="grid-item">
-              <a href="../fees.html">Fees</a>
-            </div>
-            <div class="grid-item">
-              <a href="../browse-menu.html">Browse</a>
-            </div>
-            <div class="grid-item">
-              <a href="../term_of_services.php">Term of Service</a>
-            </div>
-            <div class="grid-item">
-              <a href="../account/account.html">Account</a>
-            </div>
-            <div class="grid-item"><a href="../faq.html">FAQs</a></div>
-            <div class="grid-item">
-              <a href="../contact.html">Contact</a>
-            </div>
-            <div class="grid-item">
-              <a href="../privacy_policies.php">Privacy Policy</a>
-            </div>
-          </div>
-          <!-- Social Link -->
-          <div class="grid-item">
-            <div class="social-buttons">
-              <a href=""><i class="fab fa-instagram circle-icon"></i></a>
-              <a href=""><i class="fab fa-facebook circle-icon"></i></a>
-              <a href=""><i class="fab fa-linkedin-in circle-icon"></i></a>
-              <a href=""><i class="fab fa-twitter circle-icon"></i></a>
-            </div>
-          </div>
-        </div>
-        <hr />
-        <!-- Copyright -->
-        <p>&copy 2021 | RMIT University | Group 16</p>
-      </div>
-    </footer>
-    <!-- JavaScript -->
-    <script src="../../js/index.js"></script>
-  </body>
-</html>
