+++ templates/cart.php
@@ -1,51 +1,4 @@
-<?php
-if ( empty(session_id()) ) session_start();
-  error_reporting(E_ERROR | E_PARSE);
-  if (fopen('../php/install.php', 'r') != null) {
-      exit("'install.php' still exists! Delete it to proceed!");
-  } 
-
-  if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-    $_SESSION['postdata'] = $_POST;
-    unset($_POST);
-    header("Location: ".$_SERVER['REQUEST_URI']);
-    exit;
-  }
-    
-  if (@$_SESSION['postdata']){
-  $_POST=$_SESSION['postdata'];
-  unset($_SESSION['postdata']);
-  }
-
-  // echo '<h2>$_SESSION values</h2>';
-  // echo '<pre>';
-  // print_r($_SESSION);
-  // echo '</pre>';
-  // echo '<hr>';
-
-  // echo '<h2>$_POST values</h2>';
-  // echo '<pre>';
-  // print_r($_POST);
-  // echo '</pre>';
-  // echo '<hr>';
-
-  // unset($_SESSION['user']);
-
-  if (!isset($_SESSION['user']) && $_POST['hit-button'] == 'Order')
-  {
-    $_SESSION['visited-cart-page'] = 'already';
-    header('location: sign-up-form.php');
-  }
-
-  if (isset($_SESSION['user']) && $_POST['hit-button'] == 'Order' && isset($_SESSION['cart']))
-  {
-    header('location: thank_you.php');
-  }
-
-  if ($_POST['hit-button'] == 'Continue Shopping') {
-    header('location: ../index.php');
-  }
-?>
+<?php require '../php/cart_require.php'?>
 
 <!DOCTYPE html>
 <html lang="en">
