+++ templates/privacy_policies.php
@@ -1,10 +1,12 @@
 <?php
-if ( empty(session_id()) ) session_start();
-error_reporting(E_ERROR | E_PARSE);
+    // check session status and start session 
+    if (empty(session_id())) session_start();
 
-if (fopen('../php/install.php', 'r') != null) {
-    exit("'install.php' still exists! Delete it to proceed!");
-}
+    // detect install.php
+    error_reporting(E_ERROR | E_PARSE);
+    if (fopen('../../php/install.php', 'r') != null) {
+        exit("'install.php' still exists! Delete it to proceed!");
+    }
 ?>
 <!DOCTYPE html>
 <html lang="en">
